backend/django_app/requirements.txt
Django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.4
dj-database-url>=2.2
python-dotenv>=1.0
psycopg[binary]>=3.2
gunicorn>=22.0
groq>=0.11.0
supabase>=2.6.0
Pillow>=10.0
pytesseract>=0.3.10
pypdf>=5.0
pdf2image>=1.17
python-dateutil>=2.9
pydantic>=2.6
websockets>=14.0
